Add new pip provider to override the upstream one
The upstream pip provider hard-codes pypi.python.org as the index to
search when determining the latest available release of a package. This
means puppet won't use our mirrors on the second run of a pip package
resource using ensure => latest. This patch adds a new provider that
inherits from the upstream one and overrides the latest() method to
just make an HTML web request at our mirrors. If there is no mirror
configured it falls back to using the XMLRPC client with
pypi.python.org.

To use this, package resources currently using the 'pip' provider will
have to switch to the 'openstack_pip' provider.

Why not propose this (or something more general and with less HTML
parsing) upstream? Per discussion on a recent pull request[1], the
puppet maintainers want to rework puppet's internal HTTP API. While we
wait for that, the pip provider remains in limbo.

Why not use the pip CLI tool to find the latest version? The XMLRPC
"search" function that pip uses for 'pip search' is not a reliable way
to find the latest version of a package, which you can see by searching
for the 'keyring' package[2] (the latest version at this time is 8.7
but the latest shown is 8.2).

[1] https://github.com/puppetlabs/puppet/pull/3984
[2] https://pypi.python.org/pypi?%3Aaction=search&term=keyring&submit=search

Change-Id: I6c24c8f99fb3f879a30d21f38d1ad883f96f7937