Block pypi.org for pip mirror tests
The beaker tests for this module use iptables to block access to pypi to
ensure the mirror functionality of the custom pip provider is working.
Pip now uses pypi.org instead of pypi.python.org and pypi.python.org is
a CNAME to pypi.org so it makes more sense to block pypi.org directly.
Also add iptables output for debugging purposes.

Change-Id: I370a403aef30dc42f3d614c5882138dac4f41ec8